flask-app/requirements.txt
awsebcli==3.20.5
bidict==0.22.1
botocore==1.29.81
cachetools==5.3.0
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
distlib==0.3.6
dnspython==2.3.0
eventlet==0.33.3
filelock==3.9.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.2
fonttools==4.39.3
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
greenlet==2.0.1
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
pathspec==0.10.1
Pillow==9.5.0
platformdirs==3.1.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==305
PyYAML==5.4.1
requests==2.26.0
rsa==4.9
semantic-version==2.8.5
six==1.14.0
SQLAlchemy==1.4.43
termcolor==1.1.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.1.9
Werkzeug==2.2.2
zipp==3.10.0
